TwoGSMs_SecondCommit
This code here involves fetching a json from the server upon establishing an internet connection, and then it arranges the messages in batches according to their messageIDs, each with a corresponding reciepeint. The messages are then sent and a confirmatory test done to check if the message has successfully gone or encountered an error. This status {"Success", "Cancelled", "Failed"} is then sent back to the Server. This sketch also includes auto-reload of SMS incase the desired smsCount to be sent is greater than the remaining sms on the GSM.